This PR fixes a few code-emitting issues where device-specific information wasn't being included in IR and bitcode output.

It also incorporates Lisa's position-independent-code fix from GitHub.

+++ libraries/dsp/src/VoiceActivityDetector.cpp
@@ -157,7 +157,7 @@ namespace dsp
         /// <summary> lookup the weight for given bin number out of windowSize bins </summary>
         double getWeight(int bin)
         {
-            assert(bin < this->_weights.size());
+            assert(bin < static_cast<int>(this->_weights.size()));
             return this->_weights[bin];
         }
